+++ ui/src/KaptureApp/components/EditorSidebar/EditorSidebar.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import SplitPane from 'react-split-pane';
 
 import { Components } from './Components';
-import { Tools } from './NewTools';
+import { Tools } from './Tools';
 
 export class EditorSidebar extends Component {
   render() {
