+++ ui/src/KaptureApp/components/EditorSidebar/Tools/ApplyTool/AddAttributeForm.js
