+++ ui/src/KaptureApp/components/EditorSidebar/Tools/ApplyTool/ApplyTool.js
