+++ ui/src/KaptureApp/components/EditorSidebar/Tools/ApplyTool/ComponentSearch.js
