+++ ui/src/KaptureApp/components/EditorSidebar/Tools/ApplyTool/ToolComponentList/ToolComponent.js
