+++ ui/src/KaptureApp/components/EditorSidebar/Tools/ApplyTool/ToolComponentList/ToolComponentList.js
