+++ ui/src/KaptureApp/components/EditorSidebar/Tools/ApplyTool/ToolComponentList/index.js
