+++ ui/src/KaptureApp/components/EditorSidebar/Tools/ApplyTool/connect.js
