+++ ui/src/KaptureApp/components/EditorSidebar/Tools/ApplyTool/index.js
