+++ ui/src/KaptureApp/components/EditorSidebar/Tools/RemoveTool/RemoveTool.js
