+++ ui/src/KaptureApp/components/EditorSidebar/Tools/RemoveTool/RemoveTool.module.css
