+++ ui/src/KaptureApp/components/EditorSidebar/Tools/RemoveTool/RemoveToolOption.js
