+++ ui/src/KaptureApp/components/EditorSidebar/Tools/RemoveTool/connect.js
