+++ ui/src/KaptureApp/components/EditorSidebar/Tools/RemoveTool/index.js
