+++ ui/src/KaptureApp/components/EditorSidebar/Tools/SelectedContainers/SelectedContainers.module.css
