+++ ui/src/KaptureApp/components/EditorSidebar/Tools/SelectedContainers/index.js
