+++ ui/src/KaptureApp/components/EditorSidebar/Tools/Tools.js
