+++ ui/src/KaptureApp/components/EditorSidebar/Tools/connect.js
